src/requirements.txt
pymongo
fastapi
openai
uvicorn
python-dotenv
google-auth
google-auth-oauthlib
chromadb
googlesearch-python
trafilatura
lxml-html-clean
redis==5.0.1
pymilvus==2.3.4
gradio==4.31.0